qqbot/requirements.txt
cairosvg==2.5.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
decorator==4.4.2
future==0.18.2
idna==2.9
jsonpickle==1.3
numpy==1.18.2
pycparser==2.20
pymongo==3.10.1
python-dateutil==2.8.1
python-telegram-bot==12.5
requests==2.23.0
six==1.14.0
tornado==6.0.4
urllib3==1.25.9
matplotlib==3.0.3
pygal==2.4.0
vaderSentiment==3.3.2